Allow setting magento home environment variable
This is to account for situations where messenger is included in the codebase via symlink.

requiring shell/abstract.php doesn't work as it is in a different real path to the messenger code.

E.g. composer with symlink strategy. Magento composer installer copy strategy has a lot of drawbacks such as file mappings going haywire.

The path of least resistance is to allow MAGE_HOME to be specified when running the scripts. e.g. MAGE_HOME=/path/to/magento/public php messenger/receiver.php stop --queues test:1:d